do not kill KWJA proc when initialization is failed

+++ src/rhoknp/processors/kwja.py
@@ -57,7 +57,7 @@ class KWJA(Processor):
         return f"{self.__class__.__name__}({arg_string})"
 
     def __del__(self) -> None:
-        if self._proc is not None:
+        if hasattr(self, "_proc") and self._proc is not None:
             self._proc.kill()
 
     def is_available(self) -> bool:
